Refactor only create person to create contents
